requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.1
PyPDF2==3.0.1
mistralai==0.1.3